Rename framework shapes
