+++ how-to/customize-workspace/client/src/framework/buttons.ts
@@ -6,7 +6,7 @@ import {
 } from "@openfin/workspace-platform";
 import { checkConditions } from "./conditions";
 import { getSettings } from "./settings";
-import type { ToolbarButtonDefinition } from "./shapes";
+import type { ToolbarButtonDefinition } from "./shapes/framework-shapes";
 import { getCurrentTheme } from "./themes";
 
 let allToolbarButtons: ToolbarButtonDefinition[];
