+++ how-to/customize-workspace/client/src/framework/settings.ts
@@ -1,7 +1,7 @@
 import * as endpointProvider from "./endpoint";
 import { createLogger } from "./logger-provider";
-import type { CustomSettings } from "./shapes";
 import type { AuthProviderOptions } from "./shapes/auth-shapes";
+import type { CustomSettings } from "./shapes/framework-shapes";
 
 let settings: CustomSettings;
 let validManifestHosts: string[];
