+++ how-to/customize-workspace/client/src/framework/shapes/framework-shapes.ts
