+++ how-to/customize-workspace/client/src/framework/shapes/index.ts
@@ -3,10 +3,10 @@ export * from "./auth-shapes";
 export * from "./conditions-shapes";
 export * from "./connection-shapes";
 export * from "./endpoint-shapes";
+export * from "./framework-shapes";
 export * from "./init-options-shapes";
 export * from "./integrations-shapes";
 export * from "./lifecycle-shapes";
 export * from "./logger-shapes";
 export * from "./module-shapes";
-export * from "./shapes";
 export * from "./template-shapes";
